pyproject.toml
[project]
name = "fzi-sem-agent-orchestration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "langsmith>=0.4.49",
    "langfuse>=3.0.0",
    "marimo>=0.18.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "streampipes>=0.97.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "data*"]
